moves api,provider,store to network folder
